add: content detail skeleton & back drop is option by props
